EDA-why past analysis of fish age failed
The number of abalone has declined greatly due to over harvesting. Understanding of abalone age(adult or infant) and abalone sex is crucial for the effort of protecting abalone stability. 
This exploratory data analysis has presented why the past analysis of abalone fish age has failed, and why the analysis of sex of immature fish has failed. It suggests additional physical parameters to collect in order to successfully predict age of abalone and the sex of immature fish.